(#15699) Add methods for resolving and more adequately stubbing the forge host
This patch includes the new 'forge' method which looks up the host using the
following mechanisms:

* return the host with the 'forge' role assigned to it
* return the value of 'forge_host' in the CONFIG hash
* default to returning 'forge-acceptance.puppetlabs.lan'

This allows us multiple mechanisms to override the forge host, with a fall back
to the static 'forge-acceptance.puppetlabs.lan' host.

It also provides wrapper methodology for stubbing a host in general
`host_alias_stub` and provides a forge specific wrapper `forge_host_alias_stub`
so we can make our acceptance code more DRY.

The `forge_host_alias_stub` is responsible for stubbing a host alias and
removing the stub when complete for each wrapped test case.